Timer.h: reloj de 64 bits para todos los cronometros del proyecto (nowMs/Stopwatch/Ticker).
Timer.h centraliza el tiempo en esp_timer_get_time() (microsegundos de arranque, 64 bits, sin envuelta en ~292.000 anos): nowMs() devuelve el instante, Stopwatch mide plazos y Ticker acumula pasos periodicos. Todas las clases que usaban millis() (32 bits, envuelta cada ~49,7 dias) pasan a este reloj:

- Boot: avance (Ticker ANIM_TICK) y duracion total (Stopwatch TOTAL_MS); el while del acumulador se reduce a consume().
- Menu: parpadeos del rombo (_hold) y de las flechas de los selectores (_editBlink) anclados al entrar en la ventana/modo (sin salto de fase); holdRepeat con dos Stopwatch (_repeat/_repeatTick).
- Legend: ciclo DWELL_MS y parpadeo del rombo activo con Stopwatch _timer.
- Scroller: avance con Ticker ANIM_TICK (consume()).
- Game: _moveLast/_startMs/_gameOverMs en uint64_t con nowMs().
- Buzzer: _startMs en uint64_t; fin de duracion por resta 64 bits.
- Buttons: debounce con nowMs().

PROYECTO.md actualizado (secciones 1, 3 y nueva seccion 21).

+++ Buttons.h
@@ -135,7 +135,7 @@ private:
   // Debounce
   // ========================================================
 
-  uint32_t _buttonLast[MAX_BUTTONS];
+  uint64_t _buttonLast[MAX_BUTTONS];
   uint32_t _buttonDelay;
 };
